+++ test/account/RoleAccount.test.js
@@ -1,0 +1,39 @@
+const { ethers } = require('hardhat');
+const { expect } = require('chai');
+
+// Note that most tests related to RoleAccount are in test/access/manager/AccessManagerWithRoleAccounts.test.js
+describe('RoleAccount', function () {
+  it('should revert if deployed with address(0) access manager', async function () {
+    const factory = await ethers.getContractFactory('$RoleAccount');
+    await expect(ethers.deployContract('$RoleAccount', [ethers.ZeroAddress])).to.be.revertedWithCustomError(
+      factory,
+      'InvalidAccessManager',
+    );
+  });
+
+  it('should return the admin role if called directly', async function () {
+    const [admin] = await ethers.getSigners();
+    const manager = await ethers.deployContract('$AccessManager', [admin]);
+
+    const implementation = await ethers.deployContract('$RoleAccount', [manager]);
+
+    await expect(implementation.accessManager()).to.eventually.equal(manager);
+    await expect(implementation.roleId()).to.eventually.equal(0n);
+    await expect(implementation.$_isUnrestrictedMember(admin)).to.eventually.be.true;
+  });
+
+  it('should return the admin role if deployed via clones without immutable args', async function () {
+    const [admin] = await ethers.getSigners();
+    const manager = await ethers.deployContract('$AccessManager', [admin]);
+
+    const factory = await ethers.deployContract('$Clones');
+    const implementation = await ethers.deployContract('$RoleAccount', [manager]);
+
+    const signer = await factory.$clone.staticCall(implementation).then(address => implementation.attach(address));
+    await factory.$clone(implementation);
+
+    await expect(signer.accessManager()).to.eventually.equal(manager);
+    await expect(signer.roleId()).to.eventually.equal(0n);
+    await expect(signer.$_isUnrestrictedMember(admin)).to.eventually.be.true;
+  });
+});